% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{Fit}
\alias{Fit}
\title{Fit Function}
\usage{
Fit(
  dataset,
  model,
  batch.size = 8192,
  epochs = 1500,
  val.split = 0.2,
  remodel.dir,
  i
)
}
\arguments{
\item{dataset}{Training and test data}

\item{model}{Keras model}

\item{batch.size}{Batch size}

\item{epochs}{Number of training epochs}

\item{val.split}{Validation split}

\item{remodel.dir}{Directory that contains model files that are saved after every epoch}

\item{i}{Model number}
}
\description{
This function trains an existing deep neural network metamodel.
}
\examples{
Fit(dataset, model, batch.size, epochs, val.split, remodel.dir, i)
}
