% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{Plot Function}
\usage{
Plot(x, history)
}
\arguments{
\item{x}{Model number}

\item{history}{Training history}
}
\description{
This function generates and saves plots and data.
}
\examples{
Plot(x, history)
}
