% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{Predict}
\alias{Predict}
\title{Predict Function}
\usage{
Predict(
  bn,
  cores = parallel::detectCores()/2,
  metamodel,
  keff.cutoff = 0.9,
  mass.cutoff = 100,
  rad.cutoff = 7,
  sample.size = 1e+06,
  ext.dir
)
}
\arguments{
\item{bn}{Bayesian network object}

\item{cores}{Number of CPU cores to use for generating Bayesian network samples}

\item{metamodel}{List of deep neural network metamodels and weights}

\item{keff.cutoff}{keff cutoff value (e.g., 0.9)}

\item{mass.cutoff}{mass cutoff (grams)}

\item{rad.cutoff}{radius cutoff (cm)}

\item{sample.size}{Number of samples used to calculate risk}

\item{ext.dir}{External directory (full path)}
}
\value{
A list of Bayesian network samples with predicted keff values
}
\description{
This function samples the Bayesian network and generates keff predictions using a deep neural network metamodel.
}
