% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{Risk}
\alias{Risk}
\title{Risk Function}
\usage{
Risk(
  bn,
  code = "mcnp",
  dataset,
  dist = "gamma",
  facility,
  keff.cutoff = 0.9,
  metamodel,
  risk.pool = 100,
  sample.size = 1e+09,
  ext.dir
)
}
\arguments{
\item{bn}{Bayesian network object}

\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{dataset}{Training and test data}

\item{dist}{Truncated probability distribution (e.g., "gamma", "normal")}

\item{facility}{Facility name or building number (.csv file name)}

\item{keff.cutoff}{keff cutoff value (e.g., 0.95)}

\item{metamodel}{List of deep neural network metamodels and weights}

\item{risk.pool}{Number of times risk is calculated}

\item{sample.size}{Number of samples used to calculate risk}

\item{ext.dir}{External directory}
}
\description{
This function imports the Sample function and estimates process criticality accident risk.
}
\examples{
Risk(
  bn,
  code = "mcnp",
  dataset,
  dist = "gamma",
  facility = "facility",
  keff.cutoff = 0.5,
  metamodel,
  risk.pool = 10,
  sample.size = 1e+05,
  ext.dir = paste0(.libPaths()[1], "/criticality/data")
)
}
