% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{Tabulate}
\alias{Tabulate}
\title{Tabulate Function}
\usage{
Tabulate(code = "mcnp", data.dir, source.dir)
}
\arguments{
\item{code}{Monte Carlo radiation transport code (e.g., 'cog', 'mcnp')}

\item{data.dir}{Data directory}

\item{source.dir}{Source directory}
}
\description{
This function imports the Scale function and loads/saves training and test data.
}
\examples{
Tabulate(code, data.dir, source.dir)
}
