% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{BN}
\alias{BN}
\title{BN Function}
\usage{
BN(facility, ext.dir)
}
\arguments{
\item{facility}{Facility name or building number (.csv file name)}

\item{ext.dir}{External directory}
}
\description{
This function creates a Bayesian network from nuclear facility data.
}
\examples{
BN(
  facility = "facility",
  ext.dir = paste0(.libPaths(), "/criticality/example"))
}
