% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{BN}
\alias{BN}
\title{BN Function}
\usage{
BN(facility, data.dir)
}
\arguments{
\item{facility}{Facility name or building number (.csv file name)}

\item{data.dir}{Data directory}
}
\description{
This function creates a Bayesian network from nuclear facility data.
}
\examples{
BN(facility, data.dir)
}
