% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{BN}
\alias{BN}
\title{BN Function}
\usage{
BN(facility, dist = "gamma", ext.dir)
}
\arguments{
\item{facility}{Facility name or building number (.csv file name)}

\item{dist}{Truncated probability distribution (e.g., "gamma", "normal")}

\item{ext.dir}{External directory}
}
\description{
This function creates a Bayesian network from pre-formatted nuclear facility data.
}
\examples{
BN(
  facility = "facility",
  dist = "gamma",
  ext.dir = paste0(.libPaths(), "/criticality/data")
)
}
