% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\title{Model Function}
\usage{
Model(
  dataset,
  layers = "8192-256-256-256-256-16",
  loss = "sse",
  opt.alg = "adamax",
  learning.rate = 0.00075
)
}
\arguments{
\item{dataset}{Training and test data}

\item{layers}{String that defines the deep neural network architecture (e.g., '64-64')}

\item{loss}{Loss function}

\item{opt.alg}{Optimization algorithm}

\item{learning.rate}{Learning rate}
}
\description{
This function builds a deep neural network metamodel.
}
\examples{
Model(dataset, layers, loss, opt.alg, learning.rate)
}
