% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{NN}
\alias{NN}
\title{NN Function}
\usage{
NN(
  dataset,
  batch.size = 8192,
  ensemble.size = 5,
  epochs = 1500,
  layers = "8192-256-256-256-256-16",
  loss = "sse",
  opt.alg = "adamax",
  learning.rate = 0.00075,
  val.split = 0.2,
  replot = TRUE,
  ext.dir
)
}
\arguments{
\item{dataset}{Training and test data}

\item{batch.size}{Batch size}

\item{ensemble.size}{Number of deep neural networks to train for the ensemble}

\item{epochs}{Number of training epochs}

\item{layers}{String that defines the deep neural network architecture (e.g., '64-64')}

\item{loss}{Loss function}

\item{opt.alg}{Optimization algorithm}

\item{learning.rate}{Learning rate}

\item{val.split}{Validation split}

\item{replot}{Boolean (TRUE/FALSE) that determines if plots should be regenerated}

\item{training.dir}{Training directory}
}
\description{
This function ties the Model, Fit, Plot, and Test functions together to build, train, and test a deep neural network metamodel.
}
\examples{
NN(dataset, batch.size, ensemble.size, epochs, layers, loss, opt.alg, learning.rate, val.split, replot, training.dir)
}
