% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{Plot Function}
\usage{
Plot(i, history = NULL, plot.dir)
}
\arguments{
\item{i}{Model number}

\item{history}{Training history}

\item{plot.dir}{Plot directory (full path)}
}
\description{
This function generates and saves plots and data.
}
