% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{Risk}
\alias{Risk}
\title{Risk Function}
\usage{
Risk(
  bn,
  code = "mcnp",
  cores = parallel::detectCores()/2,
  dist = "gamma",
  evidence = TRUE,
  metamodel,
  keff.cutoff = 0.9,
  mass.cutoff = 100,
  rad.cutoff = 7,
  risk.pool = 100,
  sample.size = 1e+06,
  usl = 0.95,
  ext.dir,
  training.dir = NULL
)
}
\arguments{
\item{bn}{Bayesian network}

\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{cores}{Number of CPU cores to use for generating Bayesian network samples}

\item{dist}{Truncated probability distribution (e.g., "gamma", "normal")}

\item{evidence}{Optional conditional evidence that can be used to generate samples}

\item{metamodel}{List of deep neural network metamodels and weights}

\item{keff.cutoff}{keff cutoff value (e.g., 0.9)}

\item{mass.cutoff}{mass cutoff (grams)}

\item{rad.cutoff}{radius cutoff (cm)}

\item{risk.pool}{Number of times risk is calculated}

\item{sample.size}{Number of samples used to calculate risk}

\item{usl}{Upper subcritical limit (e.g., 0.95)}

\item{ext.dir}{External directory (full path)}

\item{training.dir}{Training directory (full path)}
}
\value{
A list of lists containing process criticality accident risk estimates and Bayesian network samples
}
\description{
This function estimates process criticality accident risk (imports Sample function).
}
\examples{

ext.dir <- paste0(tempdir(), "/criticality/extdata")
dir.create(ext.dir, recursive = TRUE, showWarnings = FALSE)

extdata <- paste0(.libPaths()[1], "/criticality/extdata")
file.copy(paste0(extdata, "/facility.csv"), ext.dir, recursive = TRUE)
file.copy(paste0(extdata, "/mcnp-dataset.RData"), ext.dir, recursive = TRUE)

config <- FALSE
try(config <- reticulate::py_config()$available)
try(if (config == TRUE) {
  Risk(
    bn = BN(
      facility.data = "facility.csv",
      ext.dir = ext.dir),
    code = "mcnp",
    cores = 1,
    evidence = TRUE,
    metamodel = NN(
      batch.size = 128,
      ensemble.size = 1,
      epochs = 10,
      layers = "256-256-16",
      replot = FALSE,
      ext.dir = ext.dir),
    keff.cutoff = 0.5,
    mass.cutoff = 100,
    rad.cutoff = 7,
    risk.pool = 10,
    sample.size = 1e+04,
    usl = 0.95,
    ext.dir = ext.dir,
    training.dir = NULL
  )
})

}
