% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{Risk}
\alias{Risk}
\title{Risk Function}
\usage{
Risk(
  bn,
  code = "mcnp",
  dist = "gamma",
  facility,
  keff.cutoff = 0.9,
  metamodel,
  risk.pool = 100,
  sample.size = 1e+09,
  ext.dir
)
}
\arguments{
\item{bn}{Bayesian network}

\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{dist}{Truncated probability distribution (e.g., "gamma", "normal")}

\item{facility}{Facility name or building number (.csv file name)}

\item{keff.cutoff}{keff cutoff value (e.g., 0.95)}

\item{metamodel}{List of deep neural network metamodels and weights}

\item{risk.pool}{Number of times risk is calculated}

\item{sample.size}{Number of samples used to calculate risk}

\item{ext.dir}{External directory}
}
\description{
This function imports the Sample function and estimates process criticality accident risk.
}
\examples{
Risk(
  bn = BN(
    facility = "facility",
    dist = "gamma",
    ext.dir = paste0(.libPaths()[1], "/criticality/data")),
  code = "mcnp",
  dist = "gamma",
  facility = "facility",
  keff.cutoff = 0.5,
  metamodel = NN(
    batch.size = 128,
    code = "mcnp",
    ensemble.size = 1,
    epochs = 10,
    layers = "8192-256-256-256-256-16",
    loss = "sse",
    opt.alg = "adamax",
    learning.rate = 0.00075,
    val.split = 0.2,
    replot = TRUE,
    verbose = TRUE,
    ext.dir = paste0(.libPaths()[1], "/criticality/data"),
    training.dir = paste0(.libPaths()[1], "/criticality/data")),
  risk.pool = 10,
  sample.size = 1e+05,
  ext.dir = paste0(.libPaths()[1], "/criticality/data")
)
unlink(paste0(.libPaths()[1], "/criticality/data/bn-data.RData"))
unlink(paste0(.libPaths()[1], "/criticality/data/risk.csv"))
temp.dir <- unlist(strsplit(tempdir(), '\\\\\\\\'))
temp.dir <- paste0(temp.dir[1:(length(temp.dir) - 1)], collapse = "/")
unlink(temp.dir, recursive = TRUE, force = TRUE)
}
