% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{Sample}
\alias{Sample}
\title{Sample Function}
\usage{
Sample(
  bn,
  code = "mcnp",
  dataset,
  keff.cutoff = 0.9,
  metamodel,
  sample.size = 1e+09,
  ext.dir = paste0(.libPaths()[1], "/criticality/data"),
  risk.dir = paste0(.libPaths()[1], "/criticality/data/risk")
)
}
\arguments{
\item{bn}{Bayesian network object}

\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{dataset}{Training and test data}

\item{keff.cutoff}{keff cutoff value (e.g., 0.9)}

\item{metamodel}{List of deep neural network metamodels and weights}

\item{sample.size}{Number of samples used to calculate risk}

\item{ext.dir}{External directory}

\item{risk.dir}{Risk directory}
}
\description{
This function samples a Bayesian network object and uses an existing deep neural network metamodel to predict keff values.
}
\examples{
Sample(
  bn,
  code = "mcnp",
  dataset,
  keff.cutoff = 0.9,
  metamodel,
  sample.size = 1e+05,
  ext.dir = paste0(.libPaths()[1], "/criticality/data"),
  risk.dir = paste0(.libPaths()[1], "/criticality/data/risk")
)
}
