% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{Scale}
\alias{Scale}
\title{Scale Function}
\usage{
Scale(code = "mcnp", output, pkg.env)
}
\arguments{
\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{output}{Processed output from Monte Carlo radiation transport code simulations}

\item{pkg.env}{R package environment}
}
\description{
This function centers, scales, and one-hot encodes variables.
}
