% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{Scale}
\alias{Scale}
\title{Scale Function}
\usage{
Scale(code = "mcnp", output, ext.dir)
}
\arguments{
\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{output}{Processed output from Monte Carlo radiation transport code simulations}

\item{ext.dir}{External directory (full path)}
}
\value{
A list of centered, scaled, and one-hot-encoded training and test data
}
\description{
This function centers, scales, and one-hot encodes variables.
}
