% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{Scale}
\alias{Scale}
\title{Scale Function}
\usage{
Scale(code = "mcnp", dataset = NULL, output)
}
\arguments{
\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{dataset}{Training and test data}

\item{output}{Processed output from Monte Carlo radiation transport code simulations}
}
\description{
This function centers, scales, and one-hot encodes variables.
}
