% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{Tabulate}
\alias{Tabulate}
\title{Tabulate Function}
\usage{
Tabulate(code = "mcnp", ext.dir = getwd())
}
\arguments{
\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{ext.dir}{External directory (full path)}
}
\description{
This function imports the Scale function and loads/saves training and test data.
}
\examples{
Tabulate(
  code = "mcnp",
  ext.dir = paste0(.libPaths()[1], "/criticality/extdata")
)
}
