% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Test}
\alias{Test}
\title{Test Function}
\usage{
Test(dataset, ensemble.size = 5, loss = "sse", ext.dir, training.dir)
}
\arguments{
\item{dataset}{Training and test data}

\item{ensemble.size}{Number of deep neural networks in the ensemble}

\item{loss}{Loss function}

\item{ext.dir}{External directory (full path)}

\item{training.dir}{Training directory (full path)}
}
\value{
A list of deep neural network weights
}
\description{
This function calculates deep neural network metamodel weights and generates keff predictions for all training and test data.
}
