% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Test}
\alias{Test}
\title{Test Function}
\usage{
Test(
  code = "mcnp",
  dataset,
  ensemble.size = 5,
  loss = "sse",
  ext.dir,
  training.dir
)
}
\arguments{
\item{code}{Monte Carlo radiation transport code (e.g., "cog", "mcnp")}

\item{dataset}{Training and test data}

\item{ensemble.size}{Number of deep neural networks in the ensemble}

\item{loss}{Loss function}

\item{ext.dir}{External directory}

\item{training.dir}{Training directory}
}
\description{
This function optimizes an existing deep neural network metamodel and generates .csv predictions for all training and test data.
}
