% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Test}
\alias{Test}
\title{Test Function}
\usage{
Test(dataset, metamodel, training.mae, val.mae, training.dir)
}
\arguments{
\item{dataset}{Training and test data}

\item{metamodel}{List of deep neural network metamodels}

\item{training.mae}{Training data mean absolute errors}

\item{val.mae}{Cross-validation data mean absolute errors}

\item{training.dir}{Training directory}
}
\description{
This function optimizes an existing deep neural network metamodel and generates .csv predictions for all training and test data.
}
\examples{
Test(
  dataset,
  metamodel,
  training.mae,
  val.mae,
  training.dir = paste0(.libPaths(), "/criticality/example"))
}
